refactor: GroupViewType 재정의
