+++ android/app/src/main/java/com/happy/friendogly/domain/model/ClubMember.kt
@@ -1,6 +1,6 @@
 package com.happy.friendogly.domain.model
 
-data class GroupMember(
+data class ClubMember(
     val id: Long,
     val name: String,
     val imageUrl: String,
